+++ R/curvesplot.R
@@ -5,8 +5,11 @@ curvesplot <- function(extendedres, xmin = 0, xmax,
                        facetby, facetby2, free.y.scales = FALSE, 
                        ncol4faceting, colorby, removelegend = FALSE,  
                         npoints = 500, line.size = 0.2, line.alpha = 1,
-                       dose_log_transfo = FALSE)
+                       dose_log_transfo = FALSE,
+                       addBMD = FALSE, BMDtype = c("zSD", "xfold"), 
+                       point.size = 1, point.alpha = 1)
 {
+  BMDtype <- match.arg(BMDtype, c("zSD", "xfold"))
   if (missing(extendedres) | !is.data.frame(extendedres))
     stop("The first argument of curvesplot must be a dataframe 
     (see ?curvesplot for details).")
@@ -139,14 +142,14 @@ curvesplot <- function(extendedres, xmin = 0, xmax,
   if (missing(colorby) & missing(facetby))
   {
     gg <- ggplot(data = curves2plot, mapping = aes_(x = quote(x), y = quote(y), group = quote(id))) +
-      geom_line(size = line.size, alpha = line.alpha) 
+      geom_line(linewidth = line.size, alpha = line.alpha) 
   } else
     # facet only
     if (missing(colorby))
     { 
       curves2plot$facetby <- rep(extendedres[, facetby], each = npoints)
       gg <- ggplot(data = curves2plot, mapping = aes_(x = quote(x), y = quote(y), group = quote(id))) +
-        geom_line(size = line.size, alpha = line.alpha) 
+        geom_line(linewidth = line.size, alpha = line.alpha) 
       # + 
       #   facet_wrap(~ facetby, scales = scales.arg) 
     } else
@@ -155,14 +158,14 @@ curvesplot <- function(extendedres, xmin = 0, xmax,
       {
         curves2plot$colorby <- rep(extendedres[, colorby], each = npoints)
         gg <- ggplot(data = curves2plot, mapping = aes_(x = quote(x), y = quote(y), group = quote(id), colour = quote(colorby))) +
-          geom_line(size = line.size, alpha = line.alpha)  
+          geom_line(linewidth = line.size, alpha = line.alpha)  
       } else
         # color and facet
       {
         curves2plot$facetby <- rep(extendedres[, facetby], each = npoints)
         curves2plot$colorby <- rep(extendedres[, colorby], each = npoints)
         gg <- ggplot(data = curves2plot, mapping = aes_(x = quote(x), y = quote(y), group = quote(id), colour = quote(colorby))) +
-          geom_line(size = line.size, alpha = line.alpha)   
+          geom_line(linewidth = line.size, alpha = line.alpha)   
       }
   if (!missing(facetby))
   {
@@ -200,5 +203,53 @@ curvesplot <- function(extendedres, xmin = 0, xmax,
     gg <- gg + labs(color = colorby)
   }
   
+  if (addBMD)
+  {
+    if (BMDtype == "zSD")
+    {  
+      if (any(!is.element(c("BMD.zSD", "BMR.zSD"), cnames)))
+      {
+        stop("To add BMD-zSD values on the curves, the first argument of curvesplot must be a dataframe
+      containing the columns BMD.zSD and BMR.zSD.")
+      }
+      BMD2plot <- data.frame(x = extendedres$BMD.zSD, y = extendedres$BMR.zSD, id = extendedres$id)
+    } else 
+    {
+      if (any(!is.element(c("BMD.xfold", "BMR.xfold"), cnames)))
+      {
+        stop("To add BMD-xfold values on the curves, the first argument of curvesplot must be a dataframe
+      containing the columns BMD.xfold and BMR.xfold.")
+      }
+      BMD2plot <- data.frame(x = extendedres$BMD.xfold, y = extendedres$BMR.xfold, id = extendedres$id)
+    }
+    
+    if (y0shift) 
+    {
+      if (scaling) 
+      {
+        BMD2plot$y <- (BMD2plot$y - extendedres$y0) / extendedres$maxychange
+      } else
+      {
+        BMD2plot$y <- (BMD2plot$y - extendedres$y0) 
+      }
+    }
+    
+    
+    if (!missing(colorby))
+    {
+      BMD2plot$colorby <- extendedres[, colorby]
+    }
+    if (!missing(facetby)) 
+    {
+      BMD2plot$facetby <- extendedres[, facetby]
+      if (!missing(facetby2)) 
+      {
+        BMD2plot$facetby2 <- extendedres[, facetby2]
+      }
+    }
+    gg <- gg + geom_point(data = BMD2plot, size = point.size, alpha = point.alpha)
+  }
+        
+
   return(gg)
 }